+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### 7.06.2018
 
+* [Rust] Code Royale - Removed is_ready as superficial (CodinGame)
 * [Rust] Code Royale - Queen building in lines (CodinGame)
 
 #### 6.06.2018